refactor: Replace payment-service with payments
