--- payment-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/paymentservice/PaymentServiceApplication.java
@@ -1,13 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.paymentservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PaymentServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PaymentServiceApplication.class, args);
-	}
-
-}
