--- payment-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/paymentservice/core/repo/PaymentRepo.java
@@ -1,9 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.paymentservice.core.repo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.github.itsAkshayDubey.eventdrivenarchitecture.paymentservice.core.entity.Payment;
-
-public interface PaymentRepo extends JpaRepository<Payment, String>{
-
-}
