--- payment-service/src/test/java/com/github/itsAkshayDubey/eventdrivenarchitecture/paymentservice/PaymentServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.paymentservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PaymentServiceApplicationTests {
-
-//	@Test
-//	void contextLoads() {
-//	}
-
-}
