+++ payments/src/main/resources/application.properties
@@ -1,0 +1,14 @@
+server.port=8085
+spring.application.name=payments
+eureka.client.service-url.defaultZone=http://localhost:8761/eureka
+
+spring.datasource.url=jdbc:h2:file:~/payments;AUTO_SERVER=true
+spring.datasource.username=root
+spring.datasource.password=root
+spring.datasource.driver-class-name=org.h2.Driver
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto=update
+spring.h2.console.enabled=true
+spring.h2.console.settings.web-allow-others=true
+
+#axon.eventhandling.processors.order-group.mode=subscribing
