+++ payments/src/test/java/com/github/itsAkshayDubey/eventdrivenarchitecture/payments/PaymentsTests.java
@@ -1,0 +1,13 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.payments;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PaymentsTests {
+
+//	@Test
+//	void contextLoads() {
+//	}
+
+}
